requirements.txt
streamlit
google-cloud-bigquery
pandas
pandas-gbq
requests
ccxt
plotly
pyarrow
db-dtypes
